+++ src/components/oppsummering/vilkårsOppsummering/faktablokk/faktablokker/UførhetFaktablokk.tsx
@@ -5,7 +5,7 @@ import { vilkårTittelFormatted } from '~features/saksoversikt/utils';
 import { useI18n } from '~lib/hooks';
 import { UførhetStatus } from '~types/Behandlingsinformasjon';
 
-import saksbehandlingMessages from '../../../../../pages/saksbehandling/steg/uførhet/uførhet-nb';
+import saksbehandlingMessages from '../../../../../pages/saksbehandling/søknadsbehandling/uførhet/uførhet-nb';
 import Vilkårsblokk from '../../VilkårsBlokk';
 import Faktablokk, { FaktaSpacing } from '../Faktablokk';
