+++ src/pages/saksbehandling/søknadsbehandling/formue/Formue.tsx
