+++ src/pages/saksbehandling/søknadsbehandling/formue/FormueComponents.tsx
