+++ src/pages/saksbehandling/søknadsbehandling/formue/formue-nb.ts
