+++ src/pages/saksbehandling/søknadsbehandling/formue/formue.module.less
