+++ src/pages/saksbehandling/søknadsbehandling/formue/utils.ts
