+++ src/pages/saksbehandling/søknadsbehandling/framdriftsindikator/SaksbehandlingFramdriftsindikator.tsx
