+++ src/pages/saksbehandling/søknadsbehandling/framdriftsindikator/saksbehandlingFramdriftsindikator-nb.ts
