+++ src/pages/saksbehandling/søknadsbehandling/sharedI18n-nb.ts
