+++ src/pages/saksbehandling/søknadsbehandling/sharedUtils.ts
