+++ src/pages/saksbehandling/søknadsbehandling/uførhet/uførhet-nb.ts
